requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
python-dotenv
sqlalchemy
aioodbc
